train/requirements.txt
pandas
tensorflow
google-cloud
google-cloud-storage
gcsfs
